Policies and Permissions (#7)
* 🔐 Add Address policies and implement 'My Addresses' route

* 🔐 Add Category policy to restrict actions to admin users

* 🔐 Add Product policy to restrict actions to moderator users

* 🔐 Add User, Auth policy to restrict actions

* 🔐 Add Coupon policy and refactor Category policy logic

* 🔐 Add Discount policy and refactor Discount controller

* 🔐 Add Order policy and refactor Controllers

* 🔐 Finishing policies and refactor Controllers

+++ app/Http/Controllers/AddressController.php
@@ -2,11 +2,11 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Address;
+use Illuminate\Http\Request;
+use App\Services\AddressService;
 use App\Http\Requests\StoreAddressRequest;
 use App\Http\Requests\UpdateAddressRequest;
-use App\Models\Address;
-use App\Services\AddressService;
-use Illuminate\Http\Request;
 
 class AddressController extends Controller
 {
@@ -17,31 +17,54 @@ class AddressController extends Controller
         $this->addressService = $addressService;
     }
 
-    public function index()
+    public function index(Request $request)
     {
-        return response()->json($this->addressService->listAll());
+        $this->authorize('viewAny', Address::class);
+
+        $user = $request->user();
+
+        $addresses = $this->addressService->listAll($user);
+
+        return response()->json($addresses, 200);
     }
 
     public function show(Address $address)
     {
-        return response()->json($this->addressService->show($address));
+        $this->authorize('view', $address);
+        
+        $address = $this->addressService->show($address);
+
+        return response()->json($address, 200);
     }
 
     public function store(StoreAddressRequest $request)
     {
-        $address = $this->addressService->create($request->validated());
+        $this->authorize('create', Address::class);
+
+        $validated = $request->validated();
+
+        $address = $this->addressService->create($validated);
+        
         return response()->json($address, 201);
     }
 
     public function update(UpdateAddressRequest $request, Address $address)
     {
-        $address = $this->addressService->update($address, $request->validated());
-        return response()->json($address);
+        $this->authorize('update', $address);
+        
+        $validated = $request->validated();
+
+        $address = $this->addressService->update($address, $validated);
+        
+        return response()->json($address, 200);
     }
 
     public function destroy(Address $address)
     {
+        $this->authorize('delete', $address);
+
         $this->addressService->delete($address);
+        
         return response()->json(null, 204);
     }
 }
